+++ client/src/components/ProjectWizard/WizardPages/ProjectTargetPoints.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { createUseStyles } from "react-jss";
 import clsx from "clsx";
-import WizardRuleInputPanels from "../WizardRuleInputPanels";
+import RuleInputPanels from "../RuleInput/RuleInputPanels";
 
 function ProjectTargetPoints(props) {
   const { rules, onInputChange, classes } = props;
@@ -12,7 +12,7 @@ function ProjectTargetPoints(props) {
         Enter the # of parking spaces you intend to build to complete the Target
         Point calculation
       </h3>
-      <WizardRuleInputPanels
+       <RuleInputPanels
         rules={rules}
         onInputChange={onInputChange}
         suppressHeader
