+++ client/src/components/ProjectWizard/WizardPages/index.js
@@ -1,9 +1,9 @@
-import ProjectDetails from './ProjectDetails'
+import ProjectDescriptions from './ProjectDescriptions'
 import ProjectUse from './ProjectUse'
 import ProjectSpecifications from './ProjectSpecifications'
 import ProjectTargetPoints from './ProjectTargetPoints'
 import ProjectMeasures from './ProjectMeasures'
 
 export {
-  ProjectDetails, ProjectUse, ProjectSpecifications, ProjectTargetPoints, ProjectMeasures
+  ProjectDescriptions, ProjectUse, ProjectSpecifications, ProjectTargetPoints, ProjectMeasures
 }
