+++ client/src/styles/_variables.scss
@@ -37,4 +37,4 @@ $color-form-lightgrey: $color-lightgrey;
 $color-button: $color-express-park-9;
 
 
-$color-highlight: #a7c539; // Color for Prev/Next Wizard buttons
+$color-highlight: #a7c539; // Color for Prev/Next  buttons
